changes for graph page
